+++ sheep/sheep.h
@@ -3,13 +3,15 @@
 // Structures
 #include "structures/basicStructures.h"
 #include "structures/polygonStructures.h"
+#include "structures/textureMap.h"
 
 // Tools
 #include "tools/firstPersonCamera.h"
 #include "tools/building.h"
-#include "tools/buildingManager.h"
+#include "tools/entityManager.h"
 #include "tools/component.h"
 #include "tools/componentHelper.h"
+#include "tools/land.h"
 
 // Building Types
 #include "tools/buildingTypes/standardBuilding.h"