+++ src/cityscape.h
@@ -2,21 +2,20 @@
 #include "../wolf/wolf.h"
 #include "../samplefw/Sample.h"
 #include "../sheep/sheep.h"
-#include "cube.h"
-#include "texturedCube.h"
 #include "randomNumberGenerator.h"
 
 #include <iostream>
 #include <glm/gtx/string_cast.hpp>
 
-// Objects
+// Objects (unused)
 #include "cube.h"
-
+#include "texturedCube.h"
 
 class Cityscape: public Sample
 {
     public:
-        Cityscape(wolf::App* pApp) : Sample(pApp, "Generated Cityscape Object") {}
+        Cityscape(wolf::App* pApp)
+            : Sample(pApp, "Generated Cityscape Object") {}
         ~Cityscape();
 
         void init() override;
@@ -29,13 +28,9 @@ class Cityscape: public Sample
     private:
         wolf::Program* mainProgram = 0;
         sheep::FirstPersonCamera* camera;
+        sheep::EntityManager* entityManager;
         GLfloat time = 0;
 
-        GLboolean isRegen = GL_FALSE;
-
-        // TODO: Remove later
-        std::vector<sheep::StandardBuilding*> standardBuildings;
-        std::vector<sheep::PolygonBuilding*> polygonBuildings;
-        TexturedCube* texturedCube1;
+        
 
 };